// apps/desktop/src/main/audio-host/wire/generated/PluginAnalysisSettings.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type PluginAnalysisSettings = { sample_rate: number, block_size: number, level_dbfs: number, start_hz: number, end_hz: number, sweep_seconds: number, tail_seconds: number, tone_hz: number, model_order: number, mid_side: boolean, linear_excitation: string, fft_size: number, processing_speed: string, ramp_start_dbfs: number, ramp_end_dbfs: number, ramp_step_db: number, ramp_seconds: number, dynamics_levels_dbfs: [number, number, number], dynamics_seconds: [number, number, number], };